Extract projects memory store to file
